refactor: driver: Move MongoSchemaBuilder to driver
